requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0
matplotlib>=3.7.0
numpy-stl>=3.0.0
trimesh>=4.0.0
pyobjc-framework-AVFoundation>=10.0  # macOS camera name detection
pynput>=1.7.6  # Global keyboard/HID event capture
